Split oversized source files by responsibility
Refactor large Kotlin files in plugin-route-analysis and plugin into
focused modules following existing patterns (e.g. extended registration
collectors). No behavior changes; public APIs preserved via delegation.

- ArmeriaKotlinDecoratorSupport → chain/scope/target helpers
- ArmeriaRouteCollector → annotated routes + service registration
- ArmeriaRouteSupport → annotation, server builder, content scan
- ArmeriaBuilderCallHeuristics → Java/Kotlin heuristics
- ArmeriaRouteExplorerPanel → tree renderer + filtering
