+++ plugin-route-analysis/src/main/kotlin/com/linecorp/intellij/plugins/armeria/explorer/ArmeriaRouteCollector.kt
@@ -4,27 +4,17 @@ import com.intellij.ide.highlighter.JavaFileType
 import com.intellij.ide.plugins.PluginManagerCore
 import com.intellij.openapi.extensions.PluginId
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.util.text.StringUtil
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.JavaPsiFacade
-import com.intellij.psi.JavaRecursiveElementWalkingVisitor
-import com.intellij.psi.PsiClass
 import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiExpression
 import com.intellij.psi.PsiJavaFile
-import com.intellij.psi.PsiLiteralExpression
 import com.intellij.psi.PsiManager
-import com.intellij.psi.PsiMethod
 import com.intellij.psi.PsiMethodCallExpression
 import com.intellij.psi.search.FileTypeIndex
 import com.intellij.psi.search.GlobalSearchScope
-import com.intellij.psi.search.searches.AnnotatedElementsSearch
-import com.intellij.psi.search.searches.ReferencesSearch
 import com.intellij.psi.util.CachedValueProvider
 import com.intellij.psi.util.CachedValuesManager
 import com.intellij.psi.util.PsiModificationTracker
-import com.intellij.psi.util.PsiTreeUtil
-import com.linecorp.intellij.plugins.armeria.message
 import org.jetbrains.kotlin.idea.KotlinFileType
 import org.jetbrains.kotlin.psi.KtFile
 
@@ -59,10 +49,21 @@ object ArmeriaRouteCollector {
         val serverBuilderOnClasspath = psiFacade.findClass(ArmeriaRouteSupport.SERVER_BUILDER_CLASS, scope) != null
         val springBootArmeriaAvailable = ArmeriaRouteSupport.isSpringBootArmeriaAvailable(psiFacade, scope)
 
-        collectAnnotatedRoutesIndexed(project, scope, routes)
-        collectServiceRegistrationsIndexed(project, scope, routes, seenServiceRegistrations)
+        ArmeriaRouteCollectorAnnotatedRoutes.collectAnnotatedRoutesIndexed(project, scope, routes)
+        ArmeriaRouteCollectorServiceRegistration.collectServiceRegistrationsIndexed(
+            project,
+            scope,
+            routes,
+            seenServiceRegistrations,
+        )
         if (!serverBuilderOnClasspath) {
-            collectServiceRegistrationsFallback(project, scope, routes, fallbackScannedFiles, seenServiceRegistrations)
+            ArmeriaRouteCollectorServiceRegistration.collectServiceRegistrationsFallback(
+                project,
+                scope,
+                routes,
+                fallbackScannedFiles,
+                seenServiceRegistrations,
+            )
         }
         if (isKotlinPluginAvailable()) {
             ArmeriaKotlinRouteCollector.collectServiceRegistrationsFallback(
@@ -148,107 +149,6 @@ object ArmeriaRouteCollector {
     private fun collectionScope(project: Project): GlobalSearchScope =
         GlobalSearchScope.projectScope(project)
 
-    private fun collectAnnotatedRoutesIndexed(
-        project: Project,
-        scope: GlobalSearchScope,
-        routes: MutableList<ArmeriaRoute>,
-    ) {
-        val psiFacade = JavaPsiFacade.getInstance(project)
-        val seenMethods = mutableSetOf<PsiMethod>()
-        for (annotationFqn in ArmeriaRouteSupport.routeAnnotations.keys) {
-            val annotationClass = psiFacade.findClass(annotationFqn, scope) ?: continue
-            AnnotatedElementsSearch.searchPsiMethods(annotationClass, scope).forEach { method ->
-                if (!seenMethods.add(method)) {
-                    return@forEach
-                }
-                addAnnotatedRouteFromMethod(method, routes)
-            }
-        }
-    }
-
-    internal fun addAnnotatedRouteFromMethod(method: PsiMethod, routes: MutableList<ArmeriaRoute>) {
-        val annotation = ArmeriaRouteSupport.findRouteAnnotation(method) ?: return
-        val containingClass = method.containingClass ?: return
-        val classPrefix =
-            ArmeriaRouteSupport.extractPrimaryPath(containingClass.getAnnotation(ArmeriaRouteSupport.PATH_PREFIX_ANNOTATION))
-        val classDecorators =
-            ArmeriaRouteSupport.extractNames(containingClass.getAnnotation(ArmeriaRouteSupport.DECORATOR_ANNOTATION))
-        val classExceptionHandlers =
-            ArmeriaRouteSupport.extractNames(containingClass.getAnnotation(ArmeriaRouteSupport.EXCEPTION_HANDLER_ANNOTATION))
-        val paths = buildList {
-            addAll(ArmeriaRouteSupport.extractPaths(annotation.first))
-            addAll(ArmeriaRouteSupport.extractPathAnnotations(method))
-        }.ifEmpty { listOf("/") }.distinct()
-        val methodDecorators =
-            classDecorators + ArmeriaRouteSupport.extractNames(method.getAnnotation(ArmeriaRouteSupport.DECORATOR_ANNOTATION))
-        val methodExceptionHandlers = classExceptionHandlers + ArmeriaRouteSupport.extractNames(
-            method.getAnnotation(ArmeriaRouteSupport.EXCEPTION_HANDLER_ANNOTATION),
-        )
-        val target = buildMethodTarget(containingClass, method)
-        val executionHints = ArmeriaTimeoutSupport.collectExecutionHints(method)
-        for (rawPath in paths) {
-            val (pathType, normalizedPath) = ArmeriaRouteSupport.parsePathType(rawPath)
-            val combinedPath = ArmeriaRouteSupport.combinePaths(classPrefix, normalizedPath)
-            routes += ArmeriaRoute.create(
-                element = method,
-                protocol = RouteProtocol.HTTP.presentableName(),
-                httpMethod = annotation.second,
-                path = combinedPath,
-                target = target,
-                routeMatch = RouteMatch.ANNOTATED_HTTP,
-                pathType = pathType,
-                decorators = methodDecorators.distinct(),
-                exceptionHandlers = methodExceptionHandlers.distinct(),
-                executionHints = executionHints,
-                contentHints = ArmeriaAnnotatedMetadataSupport.collectContentHints(method, combinedPath, pathType),
-            )
-        }
-    }
-
-    private fun collectServiceRegistrationsIndexed(
-        project: Project,
-        scope: GlobalSearchScope,
-        routes: MutableList<ArmeriaRoute>,
-        seenServiceRegistrations: MutableSet<String>,
-    ) {
-        val psiFacade = JavaPsiFacade.getInstance(project)
-        val builderClass = psiFacade.findClass(ArmeriaRouteSupport.SERVER_BUILDER_CLASS, scope) ?: return
-        for (methodName in CoreServiceRegistrationMethod.METHOD_NAMES) {
-            for (method in builderClass.findMethodsByName(methodName, false)) {
-                ReferencesSearch.search(method, scope).forEach { reference ->
-                    val call = PsiTreeUtil.getParentOfType(reference.element, PsiMethodCallExpression::class.java)
-                        ?: return@forEach
-                    if (call.methodExpression.referenceName != methodName) {
-                        return@forEach
-                    }
-                    addServiceRegistrationFromCall(call, routes, seenServiceRegistrations)
-                }
-            }
-        }
-    }
-
-    private fun collectServiceRegistrationsFallback(
-        project: Project,
-        scope: GlobalSearchScope,
-        routes: MutableList<ArmeriaRoute>,
-        fallbackScannedFiles: MutableSet<VirtualFile>,
-        seenServiceRegistrations: MutableSet<String>,
-    ) {
-        for (virtualFile in FileTypeIndex.getFiles(JavaFileType.INSTANCE, scope)) {
-            if (virtualFile in fallbackScannedFiles) {
-                continue
-            }
-            ArmeriaRouteCollectionMetrics.current()?.filesScanned?.incrementAndGet()
-            val psiFile = PsiManager.getInstance(project).findFile(virtualFile) as? PsiJavaFile ?: continue
-            if (!referencesArmeriaJavaContent(psiFile)) {
-                continue
-            }
-            fallbackScannedFiles += virtualFile
-            ArmeriaRouteCollectionMetrics.current()?.armeriaFiles?.incrementAndGet()
-            collectServiceRegistrationsFromJavaFile(psiFile, routes, seenServiceRegistrations)
-        }
-    }
-
     fun referencesArmeriaJavaContent(file: PsiJavaFile): Boolean {
         val hasArmeriaImports = file.importList
             ?.allImportStatements
@@ -261,162 +161,7 @@ object ArmeriaRouteCollector {
         return ArmeriaRouteSupport.referencesArmeriaInText(file.viewProvider.contents)
     }
 
-    internal fun collectServiceRegistrationsFromJavaFile(
-        file: PsiJavaFile,
-        routes: MutableList<ArmeriaRoute>,
-        seenServiceRegistrations: MutableSet<String>,
-    ) {
-        file.accept(object : JavaRecursiveElementWalkingVisitor() {
-            override fun visitMethodCallExpression(expression: PsiMethodCallExpression) {
-                collectServiceRegistrationFromMethodCall(expression, routes, seenServiceRegistrations)
-                ArmeriaExtendedRegistrationCollector.visitMethodCallExpression(
-                    expression,
-                    routes,
-                    seenServiceRegistrations,
-                )
-                super.visitMethodCallExpression(expression)
-            }
-        })
-    }
-
-    internal fun collectServiceRegistrationFromMethodCall(
-        expression: PsiMethodCallExpression,
-        routes: MutableList<ArmeriaRoute>,
-        seenServiceRegistrations: MutableSet<String>,
-    ) {
-        ArmeriaRouteCollectionMetrics.current()?.methodCallsVisited?.incrementAndGet()
-        val methodName = expression.methodExpression.referenceName
-        if (methodName !in CoreServiceRegistrationMethod.METHOD_NAMES) {
-            return
-        }
-        if (!ArmeriaBuilderCallHeuristics.looksLikeJavaBuilderCall(expression)) {
-            return
-        }
-        addServiceRegistrationFromCall(expression, routes, seenServiceRegistrations)
-    }
-
-    internal fun addServiceRegistrationFromCall(
-        expression: PsiMethodCallExpression,
-        routes: MutableList<ArmeriaRoute>,
-        seenServiceRegistrations: MutableSet<String>,
-    ): Boolean {
-        val registrationKey = serviceRegistrationKey(expression) ?: return false
-        val methodName = expression.methodExpression.referenceName ?: return false
-        val arguments = expression.argumentList.expressions
-        val path = extractRegistrationPath(methodName, arguments) ?: return false
-        val implementationExpression = when (CoreServiceRegistrationMethod.fromMethodName(methodName)) {
-            CoreServiceRegistrationMethod.ANNOTATED_SERVICE -> arguments.getOrNull(1) ?: arguments.getOrNull(0)
-            CoreServiceRegistrationMethod.SERVICE, CoreServiceRegistrationMethod.SERVICE_UNDER -> arguments.getOrNull(1)
-            null -> null
-        } ?: return false
-        val target = ArmeriaRouteTargetExtractor.extractTarget(implementationExpression)
-        return addServiceRegistrationRoute(
-            element = expression,
-            registrationKey = registrationKey,
-            methodName = methodName,
-            path = path,
-            target = target,
-            targetUnresolved = ArmeriaRouteTargetExtractor.isUnresolvedTarget(implementationExpression, target),
-            implementationText = implementationExpression.text,
-            argumentCount = arguments.size,
-            routes = routes,
-            seenServiceRegistrations = seenServiceRegistrations,
-        )
-    }
-
-    internal fun addServiceRegistrationRoute(
-        element: PsiElement,
-        registrationKey: String,
-        methodName: String,
-        path: String,
-        target: String,
-        targetUnresolved: Boolean,
-        implementationText: String,
-        argumentCount: Int,
-        routes: MutableList<ArmeriaRoute>,
-        seenServiceRegistrations: MutableSet<String>,
-        decorators: List<String>? = null,
-    ): Boolean {
-        if (!seenServiceRegistrations.add(registrationKey)) {
-            return false
-        }
-        val registrationMethod = CoreServiceRegistrationMethod.fromMethodName(methodName) ?: return false
-        val protocol = ArmeriaRouteTargetExtractor.detectProtocol(implementationText)
-        val routeMatch = resolveRouteMatch(registrationMethod, protocol)
-        val annotatedServiceHasPathPrefix =
-            registrationMethod == CoreServiceRegistrationMethod.ANNOTATED_SERVICE && argumentCount > 1
-        val normalizedPath = ArmeriaRouteSupport.normalizePath(path)
-        val programmaticDecorators = decorators ?: collectProgrammaticDecorators(element, normalizedPath)
-        val timeoutHints = collectBuilderTimeoutHints(element)
-        routes += ArmeriaRoute.create(
-            element = element,
-            protocol = protocol.presentableName(),
-            httpMethod = "",
-            path = normalizedPath,
-            target = target,
-            routeMatch = routeMatch,
-            targetUnresolved = targetUnresolved,
-            isDocService = protocol == RouteProtocol.DOC_SERVICE,
-            annotatedServiceHasPathPrefix = annotatedServiceHasPathPrefix,
-            decorators = programmaticDecorators,
-            timeoutHints = timeoutHints,
-        )
-        return true
-    }
-
-    private fun resolveRouteMatch(registrationMethod: CoreServiceRegistrationMethod, protocol: RouteProtocol): RouteMatch {
-        if (protocol != RouteProtocol.HTTP) {
-            return RouteMatch.NON_HTTP
-        }
-        return when (registrationMethod) {
-            CoreServiceRegistrationMethod.SERVICE -> RouteMatch.SERVICE
-            CoreServiceRegistrationMethod.ANNOTATED_SERVICE -> RouteMatch.ANNOTATED_SERVICE
-            CoreServiceRegistrationMethod.SERVICE_UNDER -> RouteMatch.SERVICE_UNDER
-        }
-    }
-
-    private fun serviceRegistrationKey(expression: PsiMethodCallExpression): String? {
-        val virtualFile = expression.containingFile?.virtualFile ?: return null
-        val methodName = expression.methodExpression.referenceName ?: return null
-        return ArmeriaRouteSupport.registrationKey(
-            virtualFile.path,
-            expression.textRange,
-            methodName,
-        )
-    }
-
-    private fun buildMethodTarget(psiClass: PsiClass, method: PsiMethod): String {
-        val className = psiClass.qualifiedName ?: psiClass.name ?: "<anonymous>"
-        return "$className#${method.name}()"
-    }
-
-    private fun extractRegistrationPath(methodName: String, arguments: Array<PsiExpression>): String? {
-        return when (CoreServiceRegistrationMethod.fromMethodName(methodName)) {
-            CoreServiceRegistrationMethod.SERVICE, CoreServiceRegistrationMethod.SERVICE_UNDER ->
-                extractString(arguments.getOrNull(0))
-            CoreServiceRegistrationMethod.ANNOTATED_SERVICE ->
-                if (arguments.size > 1) extractString(arguments.getOrNull(0)) else "/"
-            null -> null
-        }
-    }
-
-    private fun extractString(expression: PsiExpression?): String? {
-        return when (expression) {
-            null -> null
-            is PsiLiteralExpression -> expression.value as? String
-            else -> {
-                val constantValue = JavaPsiFacade.getInstance(expression.project)
-                    .constantEvaluationHelper
-                    .computeConstantExpression(expression) as? String
-                constantValue ?: expression.text.takeIf { StringUtil.isNotEmpty(it) }
-            }
-        }
-    }
-
-    private fun isKotlinPluginAvailable(): Boolean =
-        PluginManagerCore.isLoaded(KOTLIN_PLUGIN_ID)
-
-    private fun collectProgrammaticDecorators(element: PsiElement, registrationPath: String): List<String> {
+    internal fun collectProgrammaticDecorators(element: PsiElement, registrationPath: String): List<String> {
         if (element is PsiMethodCallExpression) {
             return ArmeriaDecoratorSupport.collectProgrammaticDecorators(element, registrationPath)
         }
@@ -426,7 +171,7 @@ object ArmeriaRouteCollector {
         return emptyList()
     }
 
-    private fun collectBuilderTimeoutHints(element: PsiElement): List<String> {
+    internal fun collectBuilderTimeoutHints(element: PsiElement): List<String> {
         if (element is PsiMethodCallExpression) {
             return ArmeriaTimeoutSupport.collectBuilderTimeoutHints(element)
         }
@@ -436,4 +181,6 @@ object ArmeriaRouteCollector {
         return emptyList()
     }
 
+    private fun isKotlinPluginAvailable(): Boolean =
+        PluginManagerCore.isLoaded(KOTLIN_PLUGIN_ID)
 }
